+++ src/shops/shops.service.ts
@@ -9,7 +9,7 @@ import { Repository } from 'typeorm';
 import { ResultableInterface } from '../common/interfaces';
 import { RequestUserInterface } from '../users/interfaces';
 import { UserService } from '../users/users.service';
-import { ShopCreateDto, ShopUpdateDto } from './dto/index';
+import { ShopCreateDto, ShopUpdateDto } from './dtos/index';
 import { ShopsEntity } from '../common/entities/shops.entity';
 
 @Injectable()
